rename serveClient to serve-static-assets
