+++ server/serve-static-assets.js
